Make the Chromium build script work more like the Mozilla one

+++ makecrx.sh
@@ -1,21 +1,59 @@
 #!/usr/bin/env bash
 
-# How to do a new release:
-# 
-#  1. Run: ./makecrx.sh
-#  2. It'll ask you to name the new version.
+# Build an HTTPS Everywhere .crx Chromium extension (for Chromium 17+)
+#
+# To build the current state of the tree:
+#
+#     ./makecrx.sh
+#
+# To build a particular tagged release:
+#
+#     ./makecrx.sh <version number>
+#
+# eg:
+#
+#     ./makecrx.sh chrome-2012.1.26
+#
+# To set the version in the tree to today's date and test-build:
+#
+#     ./makecrx.sh today
 
-echo -n "previous versions: "; curl -s "https://api.github.com/repos/aaronsw/https-everywhere/downloads" | grep name | head -n 3 | sed -e 's/    "name": "https-everywhere-//' | sed -e "s/.crx\",/   /" | tr '\n' ' '; echo
 
-echo -n "New version: "
-read VERSION
+if [ -n "$1" ]; then
+  if [ "$1" = today ] ; then
+    date +%Y.%m.%d | sed s/\\.0/./g > chromium/chrome-version.txt
+  else
+    BRANCH=`git branch | head -n 1 | cut -d \  -f 2-`
+    SUBDIR=checkout
+    [ -d $SUBDIR ] || mkdir $SUBDIR
+    cp -r -f -a .git $SUBDIR
+    cd $SUBDIR
+    git reset --hard "$1"
+  fi
+fi
 
+VERSION=`cat chromium/chrome-version.txt`
+
+echo "Building chrome version" $VERSION
+
+if [ -x trivial-validate.py ]; then
+	VALIDATE=./trivial-validate.py
+else
+	VALIDATE=./trivial-validate
+fi
+if $VALIDATE src/chrome/content/rules >&2
+then
+  echo Validation of included rulesets completed. >&2
+  echo >&2
+else
+  echo ERROR: Validation of rulesets failed. >&2
+  exit 1
+fi
 
 sed -e "s/VERSION/$VERSION/g" chromium/updates-master.xml > chromium/updates.xml
 
-
 [ -d pkg ] || mkdir -p pkg
-rm -rf pkg/crx
+[ -e pkg/crx ] && rm -rf pkg/crx
 cp -r chromium pkg/crx
 cp -r src/chrome/content/rules pkg/crx/
 echo 'var rule_list = [' > pkg/crx/rule_list.js
@@ -24,15 +62,19 @@ do
     echo "\"rules/$(basename $i)\"," >> pkg/crx/rule_list.js
 done
 echo '];' >> pkg/crx/rule_list.js
-sed -i "" -e "s/VERSION/$VERSION/g" pkg/crx/manifest.json
-sed -i "" -e "s/VERSION/$VERSION/g" pkg/crx/updates.xml
+sed -i -e "s/VERSION/$VERSION/g" pkg/crx/manifest.json
+sed -i -e "s/VERSION/$VERSION/g" pkg/crx/updates.xml
 
 ## from https://code.google.com/chrome/extensions/crx.html
 
 dir=pkg/crx
 key=chromium.pem
 name=pkg/crx
-crx="pkg/https-everywhere-$VERSION.crx"
+if [ "$BUILD" ] ; then
+  crx="pkg/https-everywhere-$VERSION.crx"
+else
+  crx="pkg/https-everywhere-$VERSION~pre.crx"
+fi
 pub="$name.pub"
 sig="$name.sig"
 zip="$name.zip"
@@ -61,12 +103,13 @@ sig_len_hex=$(byte_swap $(printf '%08x\n' $(ls -l "$sig" | awk '{print $5}')))
   echo "$crmagic_hex $version_hex $pub_len_hex $sig_len_hex" | xxd -r -p
   cat "$pub" "$sig" "$zip"
 ) > "$crx"
-echo "Built .crx."
+echo "Built $crx"
+#rm -rf pkg/crx
 
-python githubhelper.py $VERSION
+#python githubhelper.py $VERSION
 
-git add chromium/updates.xml
-git commit -m "release $VERSION"
-git tag -u BBB4F59D -s chrome-$VERSION -m "release $VERSION"
-git push
-git push --tags
+#git add chromium/updates.xml
+#git commit -m "release $VERSION"
+#git tag -s chrome-$VERSION -m "release $VERSION"
+#git push
+#git push --tags
